refactor how types are imported for common
