+++ tsconfig.json
@@ -13,6 +13,5 @@
     "noLib": false,
     "composite": true
   },
-  "references": [{ "path": "./packages/common/tsconfig.json" }],
   "exclude": ["node_modules", "**/__tests__/*"]
 }
